pyproject.toml
[project]
name = "tv-backtesting"
version = "0.1.0"
description = "TradingView backtesting & Jamie Coutts signal system"
requires-python = ">=3.12"
dependencies = [
    "playwright>=1.49",
    "ccxt>=4.4",
    "pandas-ta>=0.3.14b1",
    "pandas>=2.2",
    "fastapi>=0.115",
    "uvicorn>=0.34",
    "apscheduler>=3.10,<4",
    "pyotp>=2.9",
    "python-dotenv>=1.0",
    "pydantic>=2.10",
    "pydantic-settings>=2.7",
]

[project.scripts]
tv-backtest = "tv_backtesting.__main__:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/tv_backtesting"]
